убрал приведение к float

+++ app/src/main/java/ru/sibur/android/garbagecollector/MainActivity.java
@@ -16,7 +16,7 @@ public class MainActivity extends AppCompatActivity implements OnMoneyUpdateList
     AutomataThread automataThread = null;
     public final static String MONEY_KEY = "money_key";
     public final static String PREF_NAME = "my_pref";
-    public final static float MONEY_DISPLAY_COEFFICIENT = (float) 0.01;
+    public final static float MONEY_DISPLAY_COEFFICIENT = 0.01f;
 
     public void OnMoneyUpdate() {
         SharedPreferences sharedPreferences = getSharedPreferences(PREF_NAME, MODE_PRIVATE);
